Post Controller - Create Action/View Part 1
Made some changes to the html for the Create View as well as to the code for the Post Controller itself (added the ability for the post to record the date and time it was created).

Also modified the site.css file to increase the bottom margin for labels and elements with the class form-group.